pyproject.toml
[tool.poetry]
name = "avito"
version = "0.1.0"
description = ""
authors = ["Alexander Shishenko <alex@shishenko.com>"]

[tool.poetry.dependencies]
python = "^3.6"
beautifulsoup4 = "^4.7"
Pillow = "^6.0"
pytesseract = "^0.2.6"
lxml = "^4.3"
requests = {version = "^2.21",extras = ["socks"]}
scrapy = "^1.6"
scrapy-rotating-proxies = "^0.6.1"
tablib = "^0.13.0"
dataset = "^1.1"
phonenumbers = "^8.10"

[tool.poetry.dev-dependencies]
ipython = "^7.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
